updt: move initialize function to util
